Split parse/parser_states module into parser_states/ package
